Library found on GitHub.
The JSON output from www.keyboard-layout-editor.com is not very easy to decode. The create of KLE realises this and created a JS library to allow for other programmers to use KLE JSON data in their programs. However, it is inconvenient to have to run JavaScript alongside Python, so I found this Python library which does the same thing.